strategy test package is moved outside the service package.
