+++ src/main/java/com/synopsys/integration/jenkins/scan/ScanPipelineCommands.java
@@ -5,8 +5,8 @@ import com.synopsys.integration.jenkins.scan.bridge.BridgeDownloadParameters;
 import com.synopsys.integration.jenkins.scan.exception.ScannerJenkinsException;
 import com.synopsys.integration.jenkins.scan.global.ExceptionMessages;
 import com.synopsys.integration.jenkins.scan.service.bridge.BridgeDownloadParametersService;
-import com.synopsys.integration.jenkins.scan.service.scan.strategy.ScanStrategyFactory;
-import com.synopsys.integration.jenkins.scan.service.scan.strategy.ScanStrategy;
+import com.synopsys.integration.jenkins.scan.strategy.ScanStrategyFactory;
+import com.synopsys.integration.jenkins.scan.strategy.ScanStrategy;
 import hudson.FilePath;
 import hudson.model.TaskListener;
 import java.util.Map;
