+++ src/main/java/com/synopsys/integration/jenkins/scan/strategy/ScanStrategy.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.jenkins.scan.service.scan.strategy;
+package com.synopsys.integration.jenkins.scan.strategy;
 
 import com.synopsys.integration.jenkins.scan.global.enums.ScanType;
 import java.util.Map;
